separate static resources from sandbox
